Dash Ability and Effect Polish
User Story:
As a car I want my special ability to be able to have a speed dash which shoots me forward really fast whenever used so that I can break boxes and walls

Description:
Added in the ability for the car to get a speed boost when pressed. A wind particle system also plays when moving faster than a certain speed.
Added some other visuals such as indicator and break lights and the cars areal now moves. Along with skid lines whenever breaking as well