Car Weight, Event System, Fuel Lose
User Stories:
As a car I want to be the heaviest player so that I will end up sinking when in water and can weight down certain platforms
AND
As a player I want to slowly lose fuel overtime so that I always have some exploration to do, especially when my abilities are not needed for the current puzzle
AND
As a developer I want the base framework for the event system architecture to be implemented so that we can ensure common communication between all classes
Description:
Added a death function which gets called when run out of fuel that disabled the input and enables each player to do a unique death animation.
AND
The event system, currently only being used for updating the fuel UI, but setup so can add further functionality to it. For it to operate added a player id scriptable object to ensure the appropriate players component updated when called.
AND
Some other test puzzle/tutorial features not finished such as a script to build a wall and breakable starting container walls