Bat Airborne Pitch & Yaw
User Stories:
As a Player, I want the Bat to be controlled with Look input, so that the Bat can be controlled whilst airborne.

AND

As a player, I want the Bat to automatically stand back up after falling over, so that the Bat is not permanently on its side and unhelpful.

Description:
he arrow keys are no longer bound to movement. They are now bound to Look and can control the Pitch and Yaw of the Bat whilst airborne.

Also fixed an issue when the Bat lands or crash lands and is unable to stand back up.