Fixed Animation Issues
When fixing the Bat Movement's Framerate Independence in #13648fd animations were broken where the Bat would not be seen flapping its Wings when Jumping. This has now been fixed.

Also fixed an issue where the Bat would not face in the direction it was moving on the Ground. This has now been fixed.