Bat - Differentiate Between Crashing and Landing
When landing on the Ground, the game will consider the Bat to have taken fatal fall damage, but because the Bat can be Airborne, we also want the Bat to land safely on the Ground.

If the Bat makes contact with anything, the Bat is considered to have Crashed if it lands at an angle >30 degrees, otherwise it's another happy landing.