Audio API Integration
User Story:
As a Developer, I want an audio controller that handles the game's audio, so that Oppenheim has a centralised audio API for every audio feature.

An AudioController has been added for Emma's Audio Assets.

This is an Audio API attached to every PlayerController that allows Audio to be played Globally, at a certain Transform's position, or follow a specific GameObject.

While purpose-built for PlayerControllers, it can also be used to play sounds Globally and does not need to be attached to a PlayerController to work.