Fuel Objects
User Stories:
As a player I want to gain 100% of my fuel back whenever I collect my fuel resource so that I can replenish my hunger and not die
AND
As a monkey I want to be able to collect bananas to replenish health so that I can regain health lost from taking damage
AND
As a Soldier, I would like to have ammo lying around the map so I can pick them up to reload my gun.
AND
As a Bat, I want to be able to collect items for other characters, so that all other characters keep high energy levels and replenish their health.

Description:
Added Models and functionality for each players fuel