Game Analytics
User Story:
As a developer, I want to connect the game to Unity's analytics services so that we can get further insights on the game

Description:
Connected to the analytics service so tracking of certain data can occur. Note it can take up to a day for your analytics dashboard to update with any new data.