Average Player Positions
User Story:
As a player I want the camera to be located at the average position of the 2 active players, so that both players are in the screen when they’re close.

The Spring Arm component now Targets the average position of the one or two Active Players. This means that bInheritRotation is no longer supported.