Address Issues with Bat
This also **hopefully** fixes the issues with the Bat.

User Stories:
As a Bat, I want to stop moving when I die, so that I appear dead to the player.

AND

As a Bat, I want to land on my feet, so that I don’t continue bouncing and sliding around uncontrollably.

This *addresses* the issues posed by the Bat. I can't really replicate the issues as well, but in theory this should solve it.

For future reference:
- Stopping the Bat from moving when it's dead was solved with BatMovement.ForceStopAllMovement() and called in Bat.OnDeath(). This forces everything related to the Bat's movement to zero.

- The Bat landing on its feet stopped the Bat from falling flat on its face and sort of glitching forwards on angled ramps. This was solved by simply setting the angular velocity to zero once the Bat has landed and no longer Airborne.

Another issue not addressed by user stories: ** This is not necessarily fixed because it was hard to replicate. **
- When jumping, the Bat was able to control its flight in the opposite direction and looked awkward doing so. This was solved with a physics material with high friction, only allowing Airborne controls when the Bat is gliding, and making the Bat jump in the Global Up direction, rather than local which would sometimes cause it to jump backwards. It also appears that the zero angular velocity also helped with this.