Final Major Puzzle on Stage 1
Implemented a crane puzzle where a crane holds up a boat which swings.
The players need to weight one side of the boat down so that the monkey can reach the climbable object and climb up to press the switch.
once pressed the boat will lift upwards and the players can move off it and collect the blueprint, finishing the level