Seesaw puzzle Adjustments
Fixes issues with pushable boxes where with the new car mass they were unable to be pushed, so whenever dashed into they now have a force applied in the dash direction.
Enlarged the seesaw slightly so that the car is less likely to fall off of it.
Moved the balloon boxes as the seesaw was popping them, or would go above them so could not rest over the seesaw