More Bat Fixes
Fixes:
Realign only when IsAirborne() and not IsGrounded().
When colliding with anything at an impact > than 10 m/s, stop all movement.
Ignore the effect of gravity when walking off an edge.
Automatically realign itself when the Bat is stuck.

Also limited the amount of Pitch the Bat can reach. This is so that when free falling, the Bat still aligns itself.