Balloon Items now properly save
The balloon puzzle items have now been properly integrated into the save system, along with fixes to the fuel items and pushable boxes saving.

Bug fix to the setting of an items unique id so that it now does some extra checks to ensure that if the item has no prefab at all its id is still set, as this item type was not being set before, only the items which had prefabs were.