Saving Finalized and Bat Adjustments
User Story:
As a player I want all the games data which change over the course of the game, such as positions of items and the players, to be saved whenever I interact with a checkpoint so that I do not need to always start from the beginning whenever I play

Description:
Whenever fully activate a checkpoint all sections before it has there data saved and all sections after it get their data reset. As long as no bugs/issues are found it is completed.

The bat has been modified to make the gliding ability weaker so that players cannot simply use the bat and fly to the end goal, as they now do not have enough power to go up 1 entire shelf level.
Enemies, especially the flying ones may also greatly assist in this as well when they are implemented.