Implemented Fans
User Stories:
As a player I want a ventilation fan that blows all players except the car so that we have more variety and challenge in the puzzles
AND
As a car, I want to be able to block the fans force, so that other players can hide behind me to have it blocked as well

Description:
Particles showcase when a fan is blowing and any item blocks the fans force for any item behind.
A cloud of dust type particle effect appears whenever the car enters a fans interactable area.
//NOTE: not too sure of the best way to highlight to players that the characters can hide behind the car in order to traverse. See what think of the dust particles.