Car Speed and Monkey Climb Animation
User Story:
As a car, I want my movement speed to be slow relative to the other characters but still significantly faster than in the existing submissions, so that I can quickly move blocks to correct positions in a puzzle.

Description:
Increased the wheels forward friction stiffness so that the car can move faster.
Fixed the monkey climb and jump animations as their changes did not save for some reason.