Puzzle 2 and 3 of the vent stage finalized
User Story:
As a player, I want a vent scene which showcases the various functionalities of the vents so that I can experience the new mechanics

Description:
for puzzle 2 the car simply needs to block the fans force so that all players can get past.

for puzzle three, the solider needs to shoot into the fans wind so that the bullets hit the enemy. Then the monkey will be able to climb across without being killed and activate the button, disabling the fans and closing the vents.