Camera Inherit Player Rotation
User Story:
As any player, I want the camera to be rotated relative to the facing direction of the active player so that the movements feel more intuitive

The view looks weird when moving 180 degrees.

I've known about this issue since Antipede, and only now I'll try to solve it.