Non-Physics Based Seesaw
User Story:
As a player, I want the seesaw to be manipulated with scripts rather than be fully physics based, so that I have better control of the rotation of the see saw.

Keep in mind that this isn't a replica of real-world physics - we'd use a Rigidbody for that. This might be inaccurate but hopefully fine and good enough for what we're using it for.