Puzzles 4 and 5 of stage 2
User Story:
As a player, I want a vent scene which showcases the various functionalities of the vents so that I can experience the new mechanics

Description:
Puzzle 4 involves using the bat to get to the switch, which spawns in a path the other characters can use.
Puzzle 5 involves using the bat to fly down, activating the fans and the using the other characters weight to push down the elevator. From this the monkey climbs up, enables the blueprint, it gets collected and the stage ends.

The idea behind the area these two puzzles are in is that it is a small flooded storage room, which on the other side of the wall is the entire water area.