Character Fuel UI
User Story:
As a player I want the UI displayed above their character so that each player understands which UI elements belong to them.

Notes:
I changed the UI from being displayed above to under the character because it could get quite hard to see when the UI is smack-bang in the middle, rather than where the feet/wheels are. I thought this was better, but it can easily be changed with a ScreenOffset variable.

Sprint Scrum Notes because I will forget:
There were multiple issues when making this work. With the new timed delay in the Viewport Split introduced in db44bda , the player rotation mechanic failed when the three players were too far (2 active, 1 inbound), the rotating Camera wouldn't update to track the new player. This was due to a mistake in the time checks in db44bda and has since been addressed and fixed - When rotating, GameEvents.RotatePlayer would fire and reset the TimeOfLastSplit variable. It now forces the ViewportSplit to update and check if players are too far.

The actual images themselves didn't want to update their positions. Camera.WorldToScreenPoint was producing inaccurate results in a single viewport. This was solved by setting the Player Canvas' to a Constant Pixel Size. In turn, this reduced the size of Oppenheim, but this might be replaced in the future with the new Tutorial.

That was for just one Viewport, we needed to test with two. The second Viewport would show the UI somewhere offscreen. This was solved when I noticed that the UI element was in the correct position *if* the second Viewport takes up the entire single Viewport. I eventually solved it by subtracting an offset value by half the screen's width to account for the resizing of the screen.

Also added an in-game on-screen Debug Console to execute functions manually to improve my testing life.