Stage 1 Boundary Boxes
User Story:
As a player I want the puzzles design to be rigirous enough to not allow any shortcuts/cheesing so that I have to strategise with my team members to pass the puzzles.

Stage 1:
I know this isn't what the user story asks for, but boundaries are good to stop the player from dying so easily.

Stage 2: WIP