Advanced Camera Collision Behaviour
User Story:
As a Player, I want the camera to automatically swivel and rotate such that my view of my character is not obscured, so that the camera does not completely zoom in towards the character.

While Inheriting Rotation, and the player backs up against a wall, the Spring Arm will automatically swivel out of the way to ensure maximum visibility of the player.