Tutorial In Place (World) UI
Fixed an issue with Tutorial UI Colliders showing In Place text in incorrect or unintended positions.

For reference, every single gameobject, box collider, canvas, and the InPlace Group's scale was all over the place and hilariously inconsistent.

In the future please keep ALL scales set to 1 on exactly one component, and not multiple components.