Player Joining
Added the ability for a player to join in a menu before they start the actual game. The character select screen was not used as since only two players exist, them being able to set their character beforehand made little sense as they would be switching between them quite a bit throughout the puzzles.
If we had 4 players this could be added in.